move to KeepAliveFixture
